requirements-pip.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cloudpickle 
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
cytoolz==0.11.0
dask 
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.2.0
faiss-cpu
fastjsonschema==2.16.2
fonttools==4.38.0
fqdn==1.5.1
fsspec 
idna==3.4
imagecodecs 
imageio 
importlib-metadata==5.1.0
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
locket 
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
networkx 
notebook==6.5.2
notebook_shim==0.2.2
numpy 
opencv-python==4.6.0.66
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
partd 
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7
PyWavelets 
pywin32==305
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-image 
scikit-learn==1.2.0
scipy==1.9.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
split-folders==0.5.1
stack-data==0.6.2
terminado==0.17.1
threadpoolctl==3.1.0
tifffile 
tinycss2==1.2.1
toolz 
torch==1.13.1
torchaudio==0.13.1
torchvision===0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.13
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
zipp==3.11.0
